src/requirements.txt
fastapi
torch
torchvision 
pillow
prometheus_client
uvicorn 
python-multipart 
transformers 
accelerate 
bitsandbytes
sentencepiece
opencv-python 
google-cloud-vision